refactor: ladder domain으로 변경
