+++ src/main/java/nextstep/ladder/domain/Line.java
@@ -4,7 +4,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 public class Line {
     private final List<Boolean> points;
@@ -14,8 +14,9 @@ public class Line {
     }
 
     private List<Boolean> createLine(final int countOfPerson) {
-        return IntStream.rangeClosed(0, countOfPerson)
-                .mapToObj(value -> true)
+        // TODO 랜덤으로 boolean 값이 생성되도록?
+        return Stream.generate(() -> true)
+                .limit(countOfPerson)
                 .collect(Collectors.toList());
     }
